Refactor node types into nodeTypeUtils.js
