+++ src/components/jsonTransforms/usfmToSlate.js
@@ -2,38 +2,7 @@ import {transform} from "json-transforms";
 import usfmjs from "usfm-js";
 import {objectToArrayRules} from "./usfmjsStructureRules";
 import {slateRules} from "./usfmjsToSlateRules";
-
-export const nodeTypes = {
-    TEXTWRAPPER: "textWrapper",
-    P: "p",
-    ND: "nd",
-    S: "s"
-}
-
-const inlineFormattingNodeTypes = [
-    nodeTypes.ND
-]
-
-const inlineNodeTypes = [
-    nodeTypes.TEXTWRAPPER
-].concat(inlineFormattingNodeTypes)
-
-const newlineNodeTypes = [
-    nodeTypes.P,
-    nodeTypes.S
-]
-
-export function isInlineFormattingNodeType(type) {
-    return inlineFormattingNodeTypes.includes(type)
-}
-
-export function isInlineNodeType(type) {
-    return inlineNodeTypes.includes(type)
-}
-
-export function isNewlineNodeType(type) {
-    return newlineNodeTypes.includes(type)
-}
+import {nodeTypes} from "../../utils/nodeTypeUtils";
 
 export const creationStamps = {
     POST_INIT: "post_init"
